% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amel_adplot.R
\name{amel_adplot}
\alias{amel_adplot}
\title{generate ggplot boxplot for ADMIXTURE data}
\usage{
amel_adplot(popname, fill_col)
}
\arguments{
\item{popname}{string that can be 'A', 'C', or 'M'}

\item{fill_col}{hexcode '#000000'}
}
\value{
figure
}
\description{
generate ggplot boxplot for ADMIXTURE data
}
\examples{
set later 
}
