% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop2_colnames.R
\name{amel_colnames}
\alias{amel_colnames}
\title{removes non-fst columns, population tags from fst estimates, and restores colnames from fst output to population names}
\usage{
amel_colnames(popnames, popool_fst)
}
\arguments{
\item{popnames}{list of population names IN THE ORDER THEY WERE ADDED TO MPILEUP}

\item{popool_fst}{data.frame .fst output created by popoolation2 fst-sliding.pl}
}
\value{
data.frame amel_colnames_df chrom_pos
}
\description{
removes non-fst columns, population tags from fst estimates, and restores colnames from fst output to population names
}
\examples{
ex_pop2names <- c("AHB", "PAFeral", "PurdueMB")
ex_pop2fst <- data.frame(X1=c("NC_CHROM", "NC_CHROM", "NC_CHROM"), X2=c("4000", "4100", "4200"), X3=c("19", "105", "739"), X4=c("1.000", "1.000", "1.000"), X5=c("100.0", "99.7", "98.9"), X6=c("1:2=0", "1:2=0.0006", "1:2=0.98"), X7=c("1:3=0.738", "1:3=0.89", "1:3=0.0"), X8=c("2:3=0.0000005", "2:3=0.2134", "2:3=0"))
pop2_colnames(ex_pop2names, ex_pop2fst)
}
