% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gb_gbo.R
\docType{data}
\name{gb_gbo}
\alias{gb_gbo}
\title{GBOld to GB Conversion}
\format{
\subsection{\code{gb_gbold}}{

A data frame with 2 columns and 14152 rows:
\describe{
\item{GB}{new Apis mellifera GenBank ID}
\item{GBold}{old Apis mellifera GenBank ID}
}
}
}
\source{
manually generated
}
\usage{
gb_gbo
}
\description{
A dataset of conversions between old and new BeeBase Gene IDs
}
\keyword{datasets}
