% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gb_ncbi.R
\docType{data}
\name{gb_ncbi}
\alias{gb_ncbi}
\title{GB to NCBI}
\format{
\subsection{\code{gb_ncbi}}{

A data frame with 2 columns and 22085 rows:
\describe{
\item{Transcript}{Apis mellifera RNA transcript ID}
\item{Gene}{old Apis mellifera GenBank ID}
}
}
}
\source{
manually generated
}
\usage{
gb_ncbi
}
\description{
A dataset of conversions between old BeeBase gene IDs and RNA transcripts
}
\keyword{datasets}
