% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gb_ncbi_conv.R
\docType{data}
\name{gb_ncbi_conv}
\alias{gb_ncbi_conv}
\title{GB to NCBI Conversion}
\format{
\subsection{\code{gb_ncbi_conv}}{

A data frame with 3 columns and 22578 rows:
\describe{
\item{comb}{Apis mellifera GenBank ID}
\item{GB}{old Apis mellifera GenBank ID, if ID changed}
\item{Transcript}{NCBI RNA transcripts}
}
}
}
\source{
manually generated
}
\usage{
gb_ncbi_conv
}
\description{
A dataset of old and new BeeBase gene IDs and NCBI RNA transcripts
}
\keyword{datasets}
