% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admix.R
\name{hbpop_admix}
\alias{hbpop_admix}
\title{for each ancestral population, calculate average contribution and standard deviation for each sample}
\usage{
hbpop_admix(plink3)
}
\arguments{
\item{plink3}{.plink created by ADMIXTURE}
}
\description{
for each ancestral population, calculate average contribution and standard deviation for each sample
}
\examples{
set later 
}
