% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bed.R
\name{hbpop_bed}
\alias{hbpop_bed}
\title{generate figures associated with FST on gene level of US honey bee populations}
\usage{
hbpop_bed(fst_df, n_percent)
}
\arguments{
\item{fst_df}{data.frame created by amel_popfig}

\item{n_percent}{number n\% (1-100) of FST from top, e.g. n_percent=10 gives top 10\%}
}
\value{
data.table
}
\description{
generate figures associated with FST on gene level of US honey bee populations
}
\examples{
set later 
}
