% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gendiv.R
\name{hbpop_gendiv}
\alias{hbpop_gendiv}
\title{generate figures associated with gendiv of US honey bee populations}
\usage{
hbpop_gendiv(geneouts, popnames, imm_genes, GB_GBold, GB_to_NCBI)
}
\arguments{
\item{geneouts}{data.frame generated by SNPGenie}

\item{popnames}{text file with sample names in column 1 and population categories in columns 2:# (FORMERLY ID)}

\item{imm_genes}{data.frame with GB IDs, gene function, and gene pathway}

\item{GB_GBold}{data.frame of corresponding new and old GB names}

\item{GB_to_NCBI}{data.frame of conversions from NCBI rna transcripts to GB}
}
\value{
data.frame and some plots
}
\description{
generate figures associated with gendiv of US honey bee populations
}
\examples{
set later 
}
