% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_ad.R
\name{hbpop_pca}
\alias{hbpop_pca}
\title{generate PCA based on admixture of US honey bee populations}
\usage{
hbpop_pca(plink_2.bed, popnames, pca_rn.txt)
}
\arguments{
\item{plink_2.bed}{.plink created by ADMIXTURE}

\item{popnames}{list of sample names and their population}

\item{pca_rn.txt}{data.frame single column file that has stock corresponding to the row}
}
\value{
figure pca_hbpop2022_admix
}
\description{
generate PCA based on admixture of US honey bee populations
}
\examples{
set later 
}
