% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hbpop_snpdf.R
\name{hbpop_snpdf}
\alias{hbpop_snpdf}
\title{organizes dataframes used for genetic diversity analysis}
\usage{
hbpop_snpdf(snp_gen_out, popnames)
}
\arguments{
\item{snp_gen_out}{dataframe generated by SNPGenie}

\item{popnames}{text file with sample names in column 1 and population categories in columns 2:#}
}
\value{
dataframe of merged datasets
}
\description{
organizes dataframes used for genetic diversity analysis
}
