% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hbpop_snpsing.R
\name{hbpop_snpsing}
\alias{hbpop_snpsing}
\title{analyzes individual genes for differences in genetic diversity}
\usage{
hbpop_snpsing(snp_gen_out, gene)
}
\arguments{
\item{snp_gen_out}{dataframe generated by SNPGenie}

\item{gene}{character name of gene of interest (GB#)}
}
\value{
dataframe
}
\description{
analyzes individual genes for differences in genetic diversity
}
