% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immgene_list.R
\docType{data}
\name{immgene_list}
\alias{immgene_list}
\title{Immunity-linked Genes}
\format{
\subsection{\code{immgene_list}}{

A data frame with 4 columns and 596 rows:
\describe{
\item{Gene.Name}{name of gene}
\item{Pathway}{immune pathway}
\item{Transcript}{RNA transcript ID}
\item{GBID}{A. mellifera GenBank ID}
}
}
}
\source{
see raw-data; assembled from
Brutscher, L. M., Daughenbaugh, K. F., & Flenniken, M. L. (2015). Antiviral defense mechanisms in honey bees. Current opinion in insect science, 10, 71-82. https://doi.org/10.1016/j.cois.2015.04.016
Evans, J. D., Aronstein, K., Chen, Y. P., Hetru, C., Imler, J. L., Jiang, H., Kanost, M., Thompson, G. J., Zou, Z., & Hultmark, D. (2006). Immune pathways and defence mechanisms in honey bees Apis mellifera. Insect molecular biology, 15(5), 645–656. https://doi.org/10.1111/j.1365-2583.2006.00682.x
Harpur, B. A., Guarna, M. M., Huxter, E., Higo, H., Moon, K. M., Hoover, S. E., ... & Zayed, A. (2019). Integrative genomics reveals the genetics and evolution of the honey bee’s social immune system. Genome Biology and Evolution, 11(3), 937-948. https://doi.org/10.1093/gbe/evz018
}
\usage{
immgene_list
}
\description{
A table of genes and RNA-transcripts linked to A. mellifera immune function
}
\keyword{datasets}
