% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fst_remake.R
\name{pop2_fig}
\alias{pop2_fig}
\title{generate statistical analyses and figures associated with genetic differentiation between US honey bee populations}
\usage{
pop2_fig(popool2_fst)
}
\arguments{
\item{popool2_fst}{.fst created by popoolation2 fst-sliding.pl}

\item{comp}{tsv with all comparisons listed by origin (Feral/Feral, Feral/Managed, or Managed/Managed)}
}
\description{
generate statistical analyses and figures associated with genetic differentiation between US honey bee populations
}
\examples{
set later 
}
