% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popnames_mhc.R
\docType{data}
\name{popnames_mhc}
\alias{popnames_mhc}
\title{Popnames for amel.poolseq}
\format{
\subsection{\code{popnames_mhc}}{

A data frame with 3 columns and 71 rows:
\describe{
\item{Sample}{Sample name}
\item{Field1}{type of management: Ref (reference pop), Feral, or Managed}
\item{GenePathway}{name of population: Cali, KYFeral, NCFeral, PAFeral, PAManaged, PurdueMB, Russ, Sask, TX}
}
}
}
\source{
Reference sequences generated by pooling native-range A. mellifera from
Harpur, B. A., Minaei, S., Kent, C. F., & Zayed, A. (2012). Management increases genetic diversity of honey bees via admixture. Molecular Ecology, 21(18), 4414-4421.
All other pops sequenced in study by M. Carpenter
}
\usage{
popnames_mhc
}
\description{
Dataset of samples used in Carpenter's analysis
}
\keyword{datasets}
