% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amel_adplot.R
\name{amel_adplot}
\alias{amel_adplot}
\title{generate ggplot boxplot for ADMIXTURE data}
\usage{
amel_adplot(popname = "C", popdiv)
}
\arguments{
\item{popname}{string that can be 'A', 'C', or 'M'; default=C}

\item{popdiv}{character ; "Management" or "Pop"}

\item{sort.tbl}{data.table produced by ADMIXTURE}
}
\value{
figure
}
\description{
generate ggplot boxplot for ADMIXTURE data
}
