% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop2_colnames.R
\name{amel_colnames}
\alias{amel_colnames}
\title{removes non-fst columns and numeric population colnames from fst estimates; sets colnames to population names}
\usage{
amel_colnames(popnames, popool_fst)
}
\arguments{
\item{popnames}{list of sample names IN THE ORDER THEY WERE ADDED TO MPILEUP and their population}

\item{popool_fst}{data.frame .fst output created by popoolation2 fst-sliding.pl}
}
\value{
data.frame with samples as col names and '#:#=' removed from values
}
\description{
removes non-fst columns and numeric population colnames from fst estimates; sets colnames to population names
}
\examples{
ex_pop2names <- data.frame(Sample=c("KYFeral", "Cali", "Tex"), Management=c("Feral","Managed","Managed"), Population=c("AHB", "PAFeral", "PurdueMB"))
ex_pop2fst <- data.frame(X1=c("NC_037638.1", "NC_037638.1", "NC_037638.1"), X2=c("58294", "62031", "75893"), X3=c("1", "1", "2"), X4=c("1.000", "1.000", "1.000"), X5=c("12", "5", "6"), X6=c("1:2=0", "1:2=0.0006", "1:2=0.98"), X7=c("1:3=0.738", "1:3=0.89", "1:3=0.0"), X8=c("2:3=0.0000005", "2:3=0.2134", "2:3=0"))
amel_colnames(ex_pop2names, ex_pop2fst)
}
