% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop2_colnames.R
\name{amel_colnames}
\alias{amel_colnames}
\title{removes non-fst columns and numeric population colnames from fst estimates; sets colnames to population names}
\usage{
amel_colnames(popnames, popool_fst, nco)
}
\arguments{
\item{popnames}{list of sample names IN THE ORDER THEY WERE ADDED TO MPILEUP and their population}

\item{popool_fst}{data.frame .fst output created by popoolation2 fst-sliding.pl}
}
\value{
data.frame (2) popool_fst2, amel_fst_df
}
\description{
removes non-fst columns and numeric population colnames from fst estimates; sets colnames to population names
}
\examples{
ex_pop2names <- data.frame(Sample=c("AHB_3829", "Scranton-LBG", "P02"), Management=c("Feral","Feral","Managed"), Population=c("AHB", "PAFeral", "PurdueMB"))
ex_pop2fst <- data.frame(X1=c("NC_CHROM", "NC_CHROM", "NC_CHROM"), X2=c("4000", "4100", "4200"), X3=c("19", "105", "739"), X4=c("1.000", "1.000", "1.000"), X5=c("100.0", "99.7", "98.9"), X6=c("1:2=0", "1:2=0.0006", "1:2=0.98"), X7=c("1:3=0.738", "1:3=0.89", "1:3=0.0"), X8=c("2:3=0.0000005", "2:3=0.2134", "2:3=0"))
amel_colnames(ex_pop2names, ex_pop2fst)
}
