% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fst_remake.R
\name{amel_popfig}
\alias{amel_popfig}
\title{generate statistical analyses and figures associated with genetic differentiation between US honey bee populations}
\usage{
amel_popfig(popool2_fst, comp)
}
\arguments{
\item{popool2_fst}{.fst created by popoolation2 fst-sliding.pl}

\item{comp}{tsv with all comparisons listed by origin (Feral/Feral, Feral/Managed, or Managed/Managed)}
}
\value{
several plots and figures
}
\description{
generate statistical analyses and figures associated with genetic differentiation between US honey bee populations
}
\examples{
set later 
}
