% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admix.R
\name{hbpop_admix}
\alias{hbpop_admix}
\title{generate figures associated with admixture of US honey bee populations}
\usage{
hbpop_admix(plink3, temp)
}
\arguments{
\item{plink3}{.plink created by ADMIXTURE}

\item{temp}{tsv with population names}
}
\value{
figure(s)
}
\description{
generate figures associated with admixture of US honey bee populations
}
\examples{
set later 
}
