% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bed.R
\name{hbpop_bed}
\alias{hbpop_bed}
\title{generate figures associated with FST on gene level of US honey bee populations}
\usage{
hbpop_bed(fst_df)
}
\arguments{
\item{fst_df}{created by fst_remake.R}
}
\value{
data.table
}
\description{
generate figures associated with FST on gene level of US honey bee populations
}
\examples{
set later 
}
