% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gendiv.R
\name{hbpop_gendiv}
\alias{hbpop_gendiv}
\title{generate figures associated with gendiv of US honey bee populations}
\usage{
hbpop_gendiv(geneouts, id.txt, imm_genes, GB_GBold, GB_to_NCBI)
}
\arguments{
\item{geneouts}{data.frame generated by SNPGenie}

\item{id.txt}{text file with file names and whether they're feral or managed}

\item{imm_genes}{data.frame with GB IDs, gene function, and gene pathway}

\item{GB_GBold}{data.frame of corresponding new and old GB names}

\item{GB_to_NCBI}{data.frame of conversions from NCBI rna transcripts to GB}
}
\value{
data.table and some plots
}
\description{
generate figures associated with gendiv of US honey bee populations
}
\examples{
set later 
}
