% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hbpop_snpsing.R
\name{hbpop_snpsing}
\alias{hbpop_snpsing}
\title{analyzes individual genes for differences in genetic diversity}
\usage{
hbpop_snpsing(snp_gen_out, gene)
}
\arguments{
\item{snp_gen_out}{data.frame generated by SNPGenie}
}
\value{
data.frame
}
\description{
analyzes individual genes for differences in genetic diversity
}
\examples{
tbd
}
